Extract `TsIdentifierBinding`, extract type parsing its own file
