+++ crates/rslint_parser/test_data/inline/ok/ts_array_type.rast
@@ -4,7 +4,7 @@ JsModule {
     items: JsModuleItemList [
         TsTypeAlias {
             type_token: TYPE_KW@0..19 "type" [Comments("// TYPESCRIPT"), Newline("\n")] [Whitespace(" ")],
-            binding_identifier: JsIdentifierBinding {
+            binding_identifier: TsIdentifierBinding {
                 name_token: IDENT@19..21 "A" [] [Whitespace(" ")],
             },
             type_parameters: missing (required),
@@ -20,7 +20,7 @@ JsModule {
         },
         TsTypeAlias {
             type_token: TYPE_KW@32..38 "type" [Newline("\n")] [Whitespace(" ")],
-            binding_identifier: JsIdentifierBinding {
+            binding_identifier: TsIdentifierBinding {
                 name_token: IDENT@38..40 "B" [] [Whitespace(" ")],
             },
             type_parameters: missing (required),
@@ -60,7 +60,7 @@ JsModule {
   2: JS_MODULE_ITEM_LIST@0..59
     0: TS_TYPE_ALIAS@0..32
       0: TYPE_KW@0..19 "type" [Comments("// TYPESCRIPT"), Newline("\n")] [Whitespace(" ")]
-      1: JS_IDENTIFIER_BINDING@19..21
+      1: TS_IDENTIFIER_BINDING@19..21
         0: IDENT@19..21 "A" [] [Whitespace(" ")]
       2: (empty)
       3: EQ@21..23 "=" [] [Whitespace(" ")]
@@ -72,7 +72,7 @@ JsModule {
       5: SEMICOLON@31..32 ";" [] []
     1: TS_TYPE_ALIAS@32..59
       0: TYPE_KW@32..38 "type" [Newline("\n")] [Whitespace(" ")]
-      1: JS_IDENTIFIER_BINDING@38..40
+      1: TS_IDENTIFIER_BINDING@38..40
         0: IDENT@38..40 "B" [] [Whitespace(" ")]
       2: (empty)
       3: EQ@40..42 "=" [] [Whitespace(" ")]
