+++ crates/rslint_parser/test_data/inline/ok/ts_conditional_type.rast
@@ -4,7 +4,7 @@ JsModule {
     items: JsModuleItemList [
         TsTypeAlias {
             type_token: TYPE_KW@0..19 "type" [Comments("// TYPESCRIPT"), Newline("\n")] [Whitespace(" ")],
-            binding_identifier: JsIdentifierBinding {
+            binding_identifier: TsIdentifierBinding {
                 name_token: IDENT@19..21 "A" [] [Whitespace(" ")],
             },
             type_parameters: missing (required),
@@ -16,7 +16,7 @@ JsModule {
         },
         TsTypeAlias {
             type_token: TYPE_KW@30..36 "type" [Newline("\n")] [Whitespace(" ")],
-            binding_identifier: JsIdentifierBinding {
+            binding_identifier: TsIdentifierBinding {
                 name_token: IDENT@36..38 "B" [] [Whitespace(" ")],
             },
             type_parameters: missing (required),
@@ -42,7 +42,7 @@ JsModule {
         },
         TsTypeAlias {
             type_token: TYPE_KW@80..86 "type" [Newline("\n")] [Whitespace(" ")],
-            binding_identifier: JsIdentifierBinding {
+            binding_identifier: TsIdentifierBinding {
                 name_token: IDENT@86..88 "C" [] [Whitespace(" ")],
             },
             type_parameters: missing (required),
@@ -103,7 +103,7 @@ JsModule {
   2: JS_MODULE_ITEM_LIST@0..148
     0: TS_TYPE_ALIAS@0..30
       0: TYPE_KW@0..19 "type" [Comments("// TYPESCRIPT"), Newline("\n")] [Whitespace(" ")]
-      1: JS_IDENTIFIER_BINDING@19..21
+      1: TS_IDENTIFIER_BINDING@19..21
         0: IDENT@19..21 "A" [] [Whitespace(" ")]
       2: (empty)
       3: EQ@21..23 "=" [] [Whitespace(" ")]
@@ -112,7 +112,7 @@ JsModule {
       5: SEMICOLON@29..30 ";" [] []
     1: TS_TYPE_ALIAS@30..80
       0: TYPE_KW@30..36 "type" [Newline("\n")] [Whitespace(" ")]
-      1: JS_IDENTIFIER_BINDING@36..38
+      1: TS_IDENTIFIER_BINDING@36..38
         0: IDENT@36..38 "B" [] [Whitespace(" ")]
       2: (empty)
       3: EQ@38..40 "=" [] [Whitespace(" ")]
@@ -131,7 +131,7 @@ JsModule {
       5: SEMICOLON@79..80 ";" [] []
     2: TS_TYPE_ALIAS@80..148
       0: TYPE_KW@80..86 "type" [Newline("\n")] [Whitespace(" ")]
-      1: JS_IDENTIFIER_BINDING@86..88
+      1: TS_IDENTIFIER_BINDING@86..88
         0: IDENT@86..88 "C" [] [Whitespace(" ")]
       2: (empty)
       3: EQ@88..90 "=" [] [Whitespace(" ")]
