+++ crates/rslint_parser/test_data/inline/ok/ts_construct_signature_member.rast
@@ -4,7 +4,7 @@ JsModule {
     items: JsModuleItemList [
         TsTypeAlias {
             type_token: TYPE_KW@0..19 "type" [Comments("// TYPESCRIPT"), Newline("\n")] [Whitespace(" ")],
-            binding_identifier: JsIdentifierBinding {
+            binding_identifier: TsIdentifierBinding {
                 name_token: IDENT@19..21 "A" [] [Whitespace(" ")],
             },
             type_parameters: missing (required),
@@ -35,7 +35,7 @@ JsModule {
         },
         TsTypeAlias {
             type_token: TYPE_KW@42..48 "type" [Newline("\n")] [Whitespace(" ")],
-            binding_identifier: JsIdentifierBinding {
+            binding_identifier: TsIdentifierBinding {
                 name_token: IDENT@48..50 "B" [] [Whitespace(" ")],
             },
             type_parameters: missing (required),
@@ -86,7 +86,7 @@ JsModule {
         },
         TsTypeAlias {
             type_token: TYPE_KW@82..88 "type" [Newline("\n")] [Whitespace(" ")],
-            binding_identifier: JsIdentifierBinding {
+            binding_identifier: TsIdentifierBinding {
                 name_token: IDENT@88..90 "C" [] [Whitespace(" ")],
             },
             type_parameters: missing (required),
@@ -176,7 +176,7 @@ JsModule {
   2: JS_MODULE_ITEM_LIST@0..126
     0: TS_TYPE_ALIAS@0..42
       0: TYPE_KW@0..19 "type" [Comments("// TYPESCRIPT"), Newline("\n")] [Whitespace(" ")]
-      1: JS_IDENTIFIER_BINDING@19..21
+      1: TS_IDENTIFIER_BINDING@19..21
         0: IDENT@19..21 "A" [] [Whitespace(" ")]
       2: (empty)
       3: EQ@21..23 "=" [] [Whitespace(" ")]
@@ -199,7 +199,7 @@ JsModule {
       5: (empty)
     1: TS_TYPE_ALIAS@42..82
       0: TYPE_KW@42..48 "type" [Newline("\n")] [Whitespace(" ")]
-      1: JS_IDENTIFIER_BINDING@48..50
+      1: TS_IDENTIFIER_BINDING@48..50
         0: IDENT@48..50 "B" [] [Whitespace(" ")]
       2: (empty)
       3: EQ@50..52 "=" [] [Whitespace(" ")]
@@ -235,7 +235,7 @@ JsModule {
       5: (empty)
     2: TS_TYPE_ALIAS@82..126
       0: TYPE_KW@82..88 "type" [Newline("\n")] [Whitespace(" ")]
-      1: JS_IDENTIFIER_BINDING@88..90
+      1: TS_IDENTIFIER_BINDING@88..90
         0: IDENT@88..90 "C" [] [Whitespace(" ")]
       2: (empty)
       3: EQ@90..92 "=" [] [Whitespace(" ")]
