+++ crates/rslint_parser/test_data/inline/ok/ts_constructor_type.rast
@@ -4,7 +4,7 @@ JsModule {
     items: JsModuleItemList [
         TsTypeAlias {
             type_token: TYPE_KW@0..19 "type" [Comments("// TYPESCRIPT"), Newline("\n")] [Whitespace(" ")],
-            binding_identifier: JsIdentifierBinding {
+            binding_identifier: TsIdentifierBinding {
                 name_token: IDENT@19..21 "A" [] [Whitespace(" ")],
             },
             type_parameters: missing (required),
@@ -53,7 +53,7 @@ JsModule {
         },
         TsTypeAlias {
             type_token: TYPE_KW@59..65 "type" [Newline("\n")] [Whitespace(" ")],
-            binding_identifier: JsIdentifierBinding {
+            binding_identifier: TsIdentifierBinding {
                 name_token: IDENT@65..67 "B" [] [Whitespace(" ")],
             },
             type_parameters: missing (required),
@@ -102,7 +102,7 @@ JsModule {
         },
         TsTypeAlias {
             type_token: TYPE_KW@114..120 "type" [Newline("\n")] [Whitespace(" ")],
-            binding_identifier: JsIdentifierBinding {
+            binding_identifier: TsIdentifierBinding {
                 name_token: IDENT@120..122 "C" [] [Whitespace(" ")],
             },
             type_parameters: missing (required),
@@ -177,7 +177,7 @@ JsModule {
         },
         TsTypeAlias {
             type_token: TYPE_KW@156..162 "type" [Newline("\n")] [Whitespace(" ")],
-            binding_identifier: JsIdentifierBinding {
+            binding_identifier: TsIdentifierBinding {
                 name_token: IDENT@162..164 "D" [] [Whitespace(" ")],
             },
             type_parameters: missing (required),
@@ -260,7 +260,7 @@ JsModule {
   2: JS_MODULE_ITEM_LIST@0..207
     0: TS_TYPE_ALIAS@0..59
       0: TYPE_KW@0..19 "type" [Comments("// TYPESCRIPT"), Newline("\n")] [Whitespace(" ")]
-      1: JS_IDENTIFIER_BINDING@19..21
+      1: TS_IDENTIFIER_BINDING@19..21
         0: IDENT@19..21 "A" [] [Whitespace(" ")]
       2: (empty)
       3: EQ@21..23 "=" [] [Whitespace(" ")]
@@ -295,7 +295,7 @@ JsModule {
       5: SEMICOLON@58..59 ";" [] []
     1: TS_TYPE_ALIAS@59..114
       0: TYPE_KW@59..65 "type" [Newline("\n")] [Whitespace(" ")]
-      1: JS_IDENTIFIER_BINDING@65..67
+      1: TS_IDENTIFIER_BINDING@65..67
         0: IDENT@65..67 "B" [] [Whitespace(" ")]
       2: (empty)
       3: EQ@67..69 "=" [] [Whitespace(" ")]
@@ -330,7 +330,7 @@ JsModule {
       5: SEMICOLON@113..114 ";" [] []
     2: TS_TYPE_ALIAS@114..156
       0: TYPE_KW@114..120 "type" [Newline("\n")] [Whitespace(" ")]
-      1: JS_IDENTIFIER_BINDING@120..122
+      1: TS_IDENTIFIER_BINDING@120..122
         0: IDENT@120..122 "C" [] [Whitespace(" ")]
       2: (empty)
       3: EQ@122..124 "=" [] [Whitespace(" ")]
@@ -383,7 +383,7 @@ JsModule {
       5: SEMICOLON@155..156 ";" [] []
     3: TS_TYPE_ALIAS@156..207
       0: TYPE_KW@156..162 "type" [Newline("\n")] [Whitespace(" ")]
-      1: JS_IDENTIFIER_BINDING@162..164
+      1: TS_IDENTIFIER_BINDING@162..164
         0: IDENT@162..164 "D" [] [Whitespace(" ")]
       2: (empty)
       3: EQ@164..166 "=" [] [Whitespace(" ")]
