+++ crates/rslint_parser/test_data/inline/ok/ts_default_type_clause.rast
@@ -4,7 +4,7 @@ JsModule {
     items: JsModuleItemList [
         TsTypeAlias {
             type_token: TYPE_KW@0..19 "type" [Comments("// TYPESCRIPT"), Newline("\n")] [Whitespace(" ")],
-            binding_identifier: JsIdentifierBinding {
+            binding_identifier: TsIdentifierBinding {
                 name_token: IDENT@19..20 "A" [] [],
             },
             type_parameters: TsTypeParameters {
@@ -36,7 +36,7 @@ JsModule {
         },
         TsTypeAlias {
             type_token: TYPE_KW@35..41 "type" [Newline("\n")] [Whitespace(" ")],
-            binding_identifier: JsIdentifierBinding {
+            binding_identifier: TsIdentifierBinding {
                 name_token: IDENT@41..42 "B" [] [],
             },
             type_parameters: TsTypeParameters {
@@ -106,7 +106,7 @@ JsModule {
   2: JS_MODULE_ITEM_LIST@0..89
     0: TS_TYPE_ALIAS@0..35
       0: TYPE_KW@0..19 "type" [Comments("// TYPESCRIPT"), Newline("\n")] [Whitespace(" ")]
-      1: JS_IDENTIFIER_BINDING@19..20
+      1: TS_IDENTIFIER_BINDING@19..20
         0: IDENT@19..20 "A" [] []
       2: TS_TYPE_PARAMETERS@20..31
         0: L_ANGLE@20..21 "<" [] []
@@ -128,7 +128,7 @@ JsModule {
       5: SEMICOLON@34..35 ";" [] []
     1: TS_TYPE_ALIAS@35..89
       0: TYPE_KW@35..41 "type" [Newline("\n")] [Whitespace(" ")]
-      1: JS_IDENTIFIER_BINDING@41..42
+      1: TS_IDENTIFIER_BINDING@41..42
         0: IDENT@41..42 "B" [] []
       2: TS_TYPE_PARAMETERS@42..79
         0: L_ANGLE@42..43 "<" [] []
