+++ crates/rslint_parser/test_data/inline/ok/ts_import_type.rast
@@ -4,7 +4,7 @@ JsModule {
     items: JsModuleItemList [
         TsTypeAlias {
             type_token: TYPE_KW@0..19 "type" [Comments("// TYPESCRIPT"), Newline("\n")] [Whitespace(" ")],
-            binding_identifier: JsIdentifierBinding {
+            binding_identifier: TsIdentifierBinding {
                 name_token: IDENT@19..21 "A" [] [Whitespace(" ")],
             },
             type_parameters: missing (required),
@@ -21,7 +21,7 @@ JsModule {
         },
         TsTypeAlias {
             type_token: TYPE_KW@45..51 "type" [Newline("\n")] [Whitespace(" ")],
-            binding_identifier: JsIdentifierBinding {
+            binding_identifier: TsIdentifierBinding {
                 name_token: IDENT@51..53 "B" [] [Whitespace(" ")],
             },
             type_parameters: missing (required),
@@ -38,7 +38,7 @@ JsModule {
         },
         TsTypeAlias {
             type_token: TYPE_KW@70..76 "type" [Newline("\n")] [Whitespace(" ")],
-            binding_identifier: JsIdentifierBinding {
+            binding_identifier: TsIdentifierBinding {
                 name_token: IDENT@76..78 "C" [] [Whitespace(" ")],
             },
             type_parameters: missing (required),
@@ -98,7 +98,7 @@ JsModule {
   2: JS_MODULE_ITEM_LIST@0..114
     0: TS_TYPE_ALIAS@0..45
       0: TYPE_KW@0..19 "type" [Comments("// TYPESCRIPT"), Newline("\n")] [Whitespace(" ")]
-      1: JS_IDENTIFIER_BINDING@19..21
+      1: TS_IDENTIFIER_BINDING@19..21
         0: IDENT@19..21 "A" [] [Whitespace(" ")]
       2: (empty)
       3: EQ@21..23 "=" [] [Whitespace(" ")]
@@ -112,7 +112,7 @@ JsModule {
       5: SEMICOLON@44..45 ";" [] []
     1: TS_TYPE_ALIAS@45..70
       0: TYPE_KW@45..51 "type" [Newline("\n")] [Whitespace(" ")]
-      1: JS_IDENTIFIER_BINDING@51..53
+      1: TS_IDENTIFIER_BINDING@51..53
         0: IDENT@51..53 "B" [] [Whitespace(" ")]
       2: (empty)
       3: EQ@53..55 "=" [] [Whitespace(" ")]
@@ -126,7 +126,7 @@ JsModule {
       5: SEMICOLON@69..70 ";" [] []
     2: TS_TYPE_ALIAS@70..114
       0: TYPE_KW@70..76 "type" [Newline("\n")] [Whitespace(" ")]
-      1: JS_IDENTIFIER_BINDING@76..78
+      1: TS_IDENTIFIER_BINDING@76..78
         0: IDENT@76..78 "C" [] [Whitespace(" ")]
       2: (empty)
       3: EQ@78..80 "=" [] [Whitespace(" ")]
