+++ crates/rslint_parser/test_data/inline/ok/ts_inferred_type.rast
@@ -4,7 +4,7 @@ JsModule {
     items: JsModuleItemList [
         TsTypeAlias {
             type_token: TYPE_KW@0..19 "type" [Comments("// TYPESCRIPT"), Newline("\n")] [Whitespace(" ")],
-            binding_identifier: JsIdentifierBinding {
+            binding_identifier: TsIdentifierBinding {
                 name_token: IDENT@19..21 "A" [] [Whitespace(" ")],
             },
             type_parameters: missing (required),
@@ -19,7 +19,7 @@ JsModule {
         },
         TsTypeAlias {
             type_token: TYPE_KW@31..37 "type" [Newline("\n")] [Whitespace(" ")],
-            binding_identifier: JsIdentifierBinding {
+            binding_identifier: TsIdentifierBinding {
                 name_token: IDENT@37..39 "B" [] [Whitespace(" ")],
             },
             type_parameters: missing (required),
@@ -75,7 +75,7 @@ JsModule {
   2: JS_MODULE_ITEM_LIST@0..67
     0: TS_TYPE_ALIAS@0..31
       0: TYPE_KW@0..19 "type" [Comments("// TYPESCRIPT"), Newline("\n")] [Whitespace(" ")]
-      1: JS_IDENTIFIER_BINDING@19..21
+      1: TS_IDENTIFIER_BINDING@19..21
         0: IDENT@19..21 "A" [] [Whitespace(" ")]
       2: (empty)
       3: EQ@21..23 "=" [] [Whitespace(" ")]
@@ -86,7 +86,7 @@ JsModule {
       5: SEMICOLON@30..31 ";" [] []
     1: TS_TYPE_ALIAS@31..67
       0: TYPE_KW@31..37 "type" [Newline("\n")] [Whitespace(" ")]
-      1: JS_IDENTIFIER_BINDING@37..39
+      1: TS_IDENTIFIER_BINDING@37..39
         0: IDENT@37..39 "B" [] [Whitespace(" ")]
       2: (empty)
       3: EQ@39..41 "=" [] [Whitespace(" ")]
