+++ crates/rslint_parser/test_data/inline/ok/ts_literal_type.rast
@@ -4,7 +4,7 @@ JsModule {
     items: JsModuleItemList [
         TsTypeAlias {
             type_token: TYPE_KW@0..19 "type" [Comments("// TYPESCRIPT"), Newline("\n")] [Whitespace(" ")],
-            binding_identifier: JsIdentifierBinding {
+            binding_identifier: TsIdentifierBinding {
                 name_token: IDENT@19..21 "A" [] [Whitespace(" ")],
             },
             type_parameters: missing (required),
@@ -17,7 +17,7 @@ JsModule {
         },
         TsTypeAlias {
             type_token: TYPE_KW@25..31 "type" [Newline("\n")] [Whitespace(" ")],
-            binding_identifier: JsIdentifierBinding {
+            binding_identifier: TsIdentifierBinding {
                 name_token: IDENT@31..33 "B" [] [Whitespace(" ")],
             },
             type_parameters: missing (required),
@@ -30,7 +30,7 @@ JsModule {
         },
         TsTypeAlias {
             type_token: TYPE_KW@38..44 "type" [Newline("\n")] [Whitespace(" ")],
-            binding_identifier: JsIdentifierBinding {
+            binding_identifier: TsIdentifierBinding {
                 name_token: IDENT@44..46 "C" [] [Whitespace(" ")],
             },
             type_parameters: missing (required),
@@ -43,7 +43,7 @@ JsModule {
         },
         TsTypeAlias {
             type_token: TYPE_KW@51..57 "type" [Newline("\n")] [Whitespace(" ")],
-            binding_identifier: JsIdentifierBinding {
+            binding_identifier: TsIdentifierBinding {
                 name_token: IDENT@57..59 "D" [] [Whitespace(" ")],
             },
             type_parameters: missing (required),
@@ -56,7 +56,7 @@ JsModule {
         },
         TsTypeAlias {
             type_token: TYPE_KW@65..71 "type" [Newline("\n")] [Whitespace(" ")],
-            binding_identifier: JsIdentifierBinding {
+            binding_identifier: TsIdentifierBinding {
                 name_token: IDENT@71..73 "E" [] [Whitespace(" ")],
             },
             type_parameters: missing (required),
@@ -68,7 +68,7 @@ JsModule {
         },
         TsTypeAlias {
             type_token: TYPE_KW@83..89 "type" [Newline("\n")] [Whitespace(" ")],
-            binding_identifier: JsIdentifierBinding {
+            binding_identifier: TsIdentifierBinding {
                 name_token: IDENT@89..91 "F" [] [Whitespace(" ")],
             },
             type_parameters: missing (required),
@@ -80,7 +80,7 @@ JsModule {
         },
         TsTypeAlias {
             type_token: TYPE_KW@98..104 "type" [Newline("\n")] [Whitespace(" ")],
-            binding_identifier: JsIdentifierBinding {
+            binding_identifier: TsIdentifierBinding {
                 name_token: IDENT@104..106 "G" [] [Whitespace(" ")],
             },
             type_parameters: missing (required),
@@ -92,7 +92,7 @@ JsModule {
         },
         TsTypeAlias {
             type_token: TYPE_KW@114..120 "type" [Newline("\n")] [Whitespace(" ")],
-            binding_identifier: JsIdentifierBinding {
+            binding_identifier: TsIdentifierBinding {
                 name_token: IDENT@120..122 "H" [] [Whitespace(" ")],
             },
             type_parameters: missing (required),
@@ -112,7 +112,7 @@ JsModule {
   2: JS_MODULE_ITEM_LIST@0..129
     0: TS_TYPE_ALIAS@0..25
       0: TYPE_KW@0..19 "type" [Comments("// TYPESCRIPT"), Newline("\n")] [Whitespace(" ")]
-      1: JS_IDENTIFIER_BINDING@19..21
+      1: TS_IDENTIFIER_BINDING@19..21
         0: IDENT@19..21 "A" [] [Whitespace(" ")]
       2: (empty)
       3: EQ@21..23 "=" [] [Whitespace(" ")]
@@ -122,7 +122,7 @@ JsModule {
       5: SEMICOLON@24..25 ";" [] []
     1: TS_TYPE_ALIAS@25..38
       0: TYPE_KW@25..31 "type" [Newline("\n")] [Whitespace(" ")]
-      1: JS_IDENTIFIER_BINDING@31..33
+      1: TS_IDENTIFIER_BINDING@31..33
         0: IDENT@31..33 "B" [] [Whitespace(" ")]
       2: (empty)
       3: EQ@33..35 "=" [] [Whitespace(" ")]
@@ -132,7 +132,7 @@ JsModule {
       5: SEMICOLON@37..38 ";" [] []
     2: TS_TYPE_ALIAS@38..51
       0: TYPE_KW@38..44 "type" [Newline("\n")] [Whitespace(" ")]
-      1: JS_IDENTIFIER_BINDING@44..46
+      1: TS_IDENTIFIER_BINDING@44..46
         0: IDENT@44..46 "C" [] [Whitespace(" ")]
       2: (empty)
       3: EQ@46..48 "=" [] [Whitespace(" ")]
@@ -142,7 +142,7 @@ JsModule {
       5: SEMICOLON@50..51 ";" [] []
     3: TS_TYPE_ALIAS@51..65
       0: TYPE_KW@51..57 "type" [Newline("\n")] [Whitespace(" ")]
-      1: JS_IDENTIFIER_BINDING@57..59
+      1: TS_IDENTIFIER_BINDING@57..59
         0: IDENT@57..59 "D" [] [Whitespace(" ")]
       2: (empty)
       3: EQ@59..61 "=" [] [Whitespace(" ")]
@@ -152,7 +152,7 @@ JsModule {
       5: SEMICOLON@64..65 ";" [] []
     4: TS_TYPE_ALIAS@65..83
       0: TYPE_KW@65..71 "type" [Newline("\n")] [Whitespace(" ")]
-      1: JS_IDENTIFIER_BINDING@71..73
+      1: TS_IDENTIFIER_BINDING@71..73
         0: IDENT@71..73 "E" [] [Whitespace(" ")]
       2: (empty)
       3: EQ@73..75 "=" [] [Whitespace(" ")]
@@ -161,7 +161,7 @@ JsModule {
       5: SEMICOLON@82..83 ";" [] []
     5: TS_TYPE_ALIAS@83..98
       0: TYPE_KW@83..89 "type" [Newline("\n")] [Whitespace(" ")]
-      1: JS_IDENTIFIER_BINDING@89..91
+      1: TS_IDENTIFIER_BINDING@89..91
         0: IDENT@89..91 "F" [] [Whitespace(" ")]
       2: (empty)
       3: EQ@91..93 "=" [] [Whitespace(" ")]
@@ -170,7 +170,7 @@ JsModule {
       5: SEMICOLON@97..98 ";" [] []
     6: TS_TYPE_ALIAS@98..114
       0: TYPE_KW@98..104 "type" [Newline("\n")] [Whitespace(" ")]
-      1: JS_IDENTIFIER_BINDING@104..106
+      1: TS_IDENTIFIER_BINDING@104..106
         0: IDENT@104..106 "G" [] [Whitespace(" ")]
       2: (empty)
       3: EQ@106..108 "=" [] [Whitespace(" ")]
@@ -179,7 +179,7 @@ JsModule {
       5: SEMICOLON@113..114 ";" [] []
     7: TS_TYPE_ALIAS@114..129
       0: TYPE_KW@114..120 "type" [Newline("\n")] [Whitespace(" ")]
-      1: JS_IDENTIFIER_BINDING@120..122
+      1: TS_IDENTIFIER_BINDING@120..122
         0: IDENT@120..122 "H" [] [Whitespace(" ")]
       2: (empty)
       3: EQ@122..124 "=" [] [Whitespace(" ")]
