+++ crates/rslint_parser/test_data/inline/ok/ts_mapped_type.rast
@@ -4,7 +4,7 @@ JsModule {
     items: JsModuleItemList [
         TsTypeAlias {
             type_token: TYPE_KW@0..19 "type" [Comments("// TYPESCRIPT"), Newline("\n")] [Whitespace(" ")],
-            binding_identifier: JsIdentifierBinding {
+            binding_identifier: TsIdentifierBinding {
                 name_token: IDENT@19..21 "A" [] [Whitespace(" ")],
             },
             type_parameters: missing (required),
@@ -40,7 +40,7 @@ JsModule {
         },
         TsTypeAlias {
             type_token: TYPE_KW@46..52 "type" [Newline("\n")] [Whitespace(" ")],
-            binding_identifier: JsIdentifierBinding {
+            binding_identifier: TsIdentifierBinding {
                 name_token: IDENT@52..64 "OptionsFlags" [] [],
             },
             type_parameters: TsTypeParameters {
@@ -90,7 +90,7 @@ JsModule {
         },
         TsTypeAlias {
             type_token: TYPE_KW@114..120 "type" [Newline("\n")] [Whitespace(" ")],
-            binding_identifier: JsIdentifierBinding {
+            binding_identifier: TsIdentifierBinding {
                 name_token: IDENT@120..133 "CreateMutable" [] [],
             },
             type_parameters: TsTypeParameters {
@@ -156,7 +156,7 @@ JsModule {
         },
         TsTypeAlias {
             type_token: TYPE_KW@199..205 "type" [Newline("\n")] [Whitespace(" ")],
-            binding_identifier: JsIdentifierBinding {
+            binding_identifier: TsIdentifierBinding {
                 name_token: IDENT@205..213 "Concrete" [] [],
             },
             type_parameters: TsTypeParameters {
@@ -222,7 +222,7 @@ JsModule {
         },
         TsTypeAlias {
             type_token: TYPE_KW@271..277 "type" [Newline("\n")] [Whitespace(" ")],
-            binding_identifier: JsIdentifierBinding {
+            binding_identifier: TsIdentifierBinding {
                 name_token: IDENT@277..284 "Getters" [] [],
             },
             type_parameters: TsTypeParameters {
@@ -343,7 +343,7 @@ JsModule {
   2: JS_MODULE_ITEM_LIST@0..389
     0: TS_TYPE_ALIAS@0..46
       0: TYPE_KW@0..19 "type" [Comments("// TYPESCRIPT"), Newline("\n")] [Whitespace(" ")]
-      1: JS_IDENTIFIER_BINDING@19..21
+      1: TS_IDENTIFIER_BINDING@19..21
         0: IDENT@19..21 "A" [] [Whitespace(" ")]
       2: (empty)
       3: EQ@21..23 "=" [] [Whitespace(" ")]
@@ -371,7 +371,7 @@ JsModule {
       5: (empty)
     1: TS_TYPE_ALIAS@46..114
       0: TYPE_KW@46..52 "type" [Newline("\n")] [Whitespace(" ")]
-      1: JS_IDENTIFIER_BINDING@52..64
+      1: TS_IDENTIFIER_BINDING@52..64
         0: IDENT@52..64 "OptionsFlags" [] []
       2: TS_TYPE_PARAMETERS@64..71
         0: L_ANGLE@64..65 "<" [] []
@@ -408,7 +408,7 @@ JsModule {
       5: SEMICOLON@113..114 ";" [] []
     2: TS_TYPE_ALIAS@114..199
       0: TYPE_KW@114..120 "type" [Newline("\n")] [Whitespace(" ")]
-      1: JS_IDENTIFIER_BINDING@120..133
+      1: TS_IDENTIFIER_BINDING@120..133
         0: IDENT@120..133 "CreateMutable" [] []
       2: TS_TYPE_PARAMETERS@133..140
         0: L_ANGLE@133..134 "<" [] []
@@ -456,7 +456,7 @@ JsModule {
       5: SEMICOLON@198..199 ";" [] []
     3: TS_TYPE_ALIAS@199..271
       0: TYPE_KW@199..205 "type" [Newline("\n")] [Whitespace(" ")]
-      1: JS_IDENTIFIER_BINDING@205..213
+      1: TS_IDENTIFIER_BINDING@205..213
         0: IDENT@205..213 "Concrete" [] []
       2: TS_TYPE_PARAMETERS@213..220
         0: L_ANGLE@213..214 "<" [] []
@@ -504,7 +504,7 @@ JsModule {
       5: SEMICOLON@270..271 ";" [] []
     4: TS_TYPE_ALIAS@271..389
       0: TYPE_KW@271..277 "type" [Newline("\n")] [Whitespace(" ")]
-      1: JS_IDENTIFIER_BINDING@277..284
+      1: TS_IDENTIFIER_BINDING@277..284
         0: IDENT@277..284 "Getters" [] []
       2: TS_TYPE_PARAMETERS@284..291
         0: L_ANGLE@284..285 "<" [] []
