+++ crates/rslint_parser/test_data/inline/ok/ts_property_or_method_signature_member.rast
@@ -4,7 +4,7 @@ JsModule {
     items: JsModuleItemList [
         TsTypeAlias {
             type_token: TYPE_KW@0..19 "type" [Comments("// TYPESCRIPT"), Newline("\n")] [Whitespace(" ")],
-            binding_identifier: JsIdentifierBinding {
+            binding_identifier: TsIdentifierBinding {
                 name_token: IDENT@19..21 "A" [] [Whitespace(" ")],
             },
             type_parameters: missing (required),
@@ -51,7 +51,7 @@ JsModule {
         },
         TsTypeAlias {
             type_token: TYPE_KW@51..57 "type" [Newline("\n")] [Whitespace(" ")],
-            binding_identifier: JsIdentifierBinding {
+            binding_identifier: TsIdentifierBinding {
                 name_token: IDENT@57..59 "B" [] [Whitespace(" ")],
             },
             type_parameters: missing (required),
@@ -98,7 +98,7 @@ JsModule {
         },
         TsTypeAlias {
             type_token: TYPE_KW@87..93 "type" [Newline("\n")] [Whitespace(" ")],
-            binding_identifier: JsIdentifierBinding {
+            binding_identifier: TsIdentifierBinding {
                 name_token: IDENT@93..95 "C" [] [Whitespace(" ")],
             },
             type_parameters: missing (required),
@@ -170,7 +170,7 @@ JsModule {
         },
         TsTypeAlias {
             type_token: TYPE_KW@140..146 "type" [Newline("\n")] [Whitespace(" ")],
-            binding_identifier: JsIdentifierBinding {
+            binding_identifier: TsIdentifierBinding {
                 name_token: IDENT@146..148 "D" [] [Whitespace(" ")],
             },
             type_parameters: missing (required),
@@ -212,7 +212,7 @@ JsModule {
         },
         TsTypeAlias {
             type_token: TYPE_KW@190..196 "type" [Newline("\n")] [Whitespace(" ")],
-            binding_identifier: JsIdentifierBinding {
+            binding_identifier: TsIdentifierBinding {
                 name_token: IDENT@196..198 "E" [] [Whitespace(" ")],
             },
             type_parameters: missing (required),
@@ -305,7 +305,7 @@ JsModule {
   2: JS_MODULE_ITEM_LIST@0..230
     0: TS_TYPE_ALIAS@0..51
       0: TYPE_KW@0..19 "type" [Comments("// TYPESCRIPT"), Newline("\n")] [Whitespace(" ")]
-      1: JS_IDENTIFIER_BINDING@19..21
+      1: TS_IDENTIFIER_BINDING@19..21
         0: IDENT@19..21 "A" [] [Whitespace(" ")]
       2: (empty)
       3: EQ@21..23 "=" [] [Whitespace(" ")]
@@ -339,7 +339,7 @@ JsModule {
       5: (empty)
     1: TS_TYPE_ALIAS@51..87
       0: TYPE_KW@51..57 "type" [Newline("\n")] [Whitespace(" ")]
-      1: JS_IDENTIFIER_BINDING@57..59
+      1: TS_IDENTIFIER_BINDING@57..59
         0: IDENT@57..59 "B" [] [Whitespace(" ")]
       2: (empty)
       3: EQ@59..61 "=" [] [Whitespace(" ")]
@@ -373,7 +373,7 @@ JsModule {
       5: (empty)
     2: TS_TYPE_ALIAS@87..140
       0: TYPE_KW@87..93 "type" [Newline("\n")] [Whitespace(" ")]
-      1: JS_IDENTIFIER_BINDING@93..95
+      1: TS_IDENTIFIER_BINDING@93..95
         0: IDENT@93..95 "C" [] [Whitespace(" ")]
       2: (empty)
       3: EQ@95..97 "=" [] [Whitespace(" ")]
@@ -423,7 +423,7 @@ JsModule {
       5: (empty)
     3: TS_TYPE_ALIAS@140..190
       0: TYPE_KW@140..146 "type" [Newline("\n")] [Whitespace(" ")]
-      1: JS_IDENTIFIER_BINDING@146..148
+      1: TS_IDENTIFIER_BINDING@146..148
         0: IDENT@146..148 "D" [] [Whitespace(" ")]
       2: (empty)
       3: EQ@148..150 "=" [] [Whitespace(" ")]
@@ -453,7 +453,7 @@ JsModule {
       5: (empty)
     4: TS_TYPE_ALIAS@190..230
       0: TYPE_KW@190..196 "type" [Newline("\n")] [Whitespace(" ")]
-      1: JS_IDENTIFIER_BINDING@196..198
+      1: TS_IDENTIFIER_BINDING@196..198
         0: IDENT@196..198 "E" [] [Whitespace(" ")]
       2: (empty)
       3: EQ@198..200 "=" [] [Whitespace(" ")]
