+++ crates/rslint_parser/test_data/inline/ok/ts_reference_type.rast
@@ -4,7 +4,7 @@ JsModule {
     items: JsModuleItemList [
         TsTypeAlias {
             type_token: TYPE_KW@0..19 "type" [Comments("// TYPESCRIPT"), Newline("\n")] [Whitespace(" ")],
-            binding_identifier: JsIdentifierBinding {
+            binding_identifier: TsIdentifierBinding {
                 name_token: IDENT@19..21 "A" [] [Whitespace(" ")],
             },
             type_parameters: missing (required),
@@ -16,7 +16,7 @@ JsModule {
         },
         TsTypeAlias {
             type_token: TYPE_KW@30..36 "type" [Newline("\n")] [Whitespace(" ")],
-            binding_identifier: JsIdentifierBinding {
+            binding_identifier: TsIdentifierBinding {
                 name_token: IDENT@36..38 "B" [] [Whitespace(" ")],
             },
             type_parameters: missing (required),
@@ -28,7 +28,7 @@ JsModule {
         },
         TsTypeAlias {
             type_token: TYPE_KW@47..53 "type" [Newline("\n")] [Whitespace(" ")],
-            binding_identifier: JsIdentifierBinding {
+            binding_identifier: TsIdentifierBinding {
                 name_token: IDENT@53..55 "C" [] [Whitespace(" ")],
             },
             type_parameters: missing (required),
@@ -43,7 +43,7 @@ JsModule {
         },
         TsTypeAlias {
             type_token: TYPE_KW@59..65 "type" [Newline("\n")] [Whitespace(" ")],
-            binding_identifier: JsIdentifierBinding {
+            binding_identifier: TsIdentifierBinding {
                 name_token: IDENT@65..67 "D" [] [Whitespace(" ")],
             },
             type_parameters: missing (required),
@@ -64,7 +64,7 @@ JsModule {
         },
         TsTypeAlias {
             type_token: TYPE_KW@73..79 "type" [Newline("\n")] [Whitespace(" ")],
-            binding_identifier: JsIdentifierBinding {
+            binding_identifier: TsIdentifierBinding {
                 name_token: IDENT@79..81 "E" [] [Whitespace(" ")],
             },
             type_parameters: missing (required),
@@ -105,7 +105,7 @@ JsModule {
   2: JS_MODULE_ITEM_LIST@0..91
     0: TS_TYPE_ALIAS@0..30
       0: TYPE_KW@0..19 "type" [Comments("// TYPESCRIPT"), Newline("\n")] [Whitespace(" ")]
-      1: JS_IDENTIFIER_BINDING@19..21
+      1: TS_IDENTIFIER_BINDING@19..21
         0: IDENT@19..21 "A" [] [Whitespace(" ")]
       2: (empty)
       3: EQ@21..23 "=" [] [Whitespace(" ")]
@@ -114,7 +114,7 @@ JsModule {
       5: SEMICOLON@29..30 ";" [] []
     1: TS_TYPE_ALIAS@30..47
       0: TYPE_KW@30..36 "type" [Newline("\n")] [Whitespace(" ")]
-      1: JS_IDENTIFIER_BINDING@36..38
+      1: TS_IDENTIFIER_BINDING@36..38
         0: IDENT@36..38 "B" [] [Whitespace(" ")]
       2: (empty)
       3: EQ@38..40 "=" [] [Whitespace(" ")]
@@ -123,7 +123,7 @@ JsModule {
       5: SEMICOLON@46..47 ";" [] []
     2: TS_TYPE_ALIAS@47..59
       0: TYPE_KW@47..53 "type" [Newline("\n")] [Whitespace(" ")]
-      1: JS_IDENTIFIER_BINDING@53..55
+      1: TS_IDENTIFIER_BINDING@53..55
         0: IDENT@53..55 "C" [] [Whitespace(" ")]
       2: (empty)
       3: EQ@55..57 "=" [] [Whitespace(" ")]
@@ -134,7 +134,7 @@ JsModule {
       5: SEMICOLON@58..59 ";" [] []
     3: TS_TYPE_ALIAS@59..73
       0: TYPE_KW@59..65 "type" [Newline("\n")] [Whitespace(" ")]
-      1: JS_IDENTIFIER_BINDING@65..67
+      1: TS_IDENTIFIER_BINDING@65..67
         0: IDENT@65..67 "D" [] [Whitespace(" ")]
       2: (empty)
       3: EQ@67..69 "=" [] [Whitespace(" ")]
@@ -149,7 +149,7 @@ JsModule {
       5: SEMICOLON@72..73 ";" [] []
     4: TS_TYPE_ALIAS@73..91
       0: TYPE_KW@73..79 "type" [Newline("\n")] [Whitespace(" ")]
-      1: JS_IDENTIFIER_BINDING@79..81
+      1: TS_IDENTIFIER_BINDING@79..81
         0: IDENT@79..81 "E" [] [Whitespace(" ")]
       2: (empty)
       3: EQ@81..83 "=" [] [Whitespace(" ")]
