+++ crates/rslint_parser/test_data/inline/ok/ts_template_literal_type.rast
@@ -4,7 +4,7 @@ JsModule {
     items: JsModuleItemList [
         TsTypeAlias {
             type_token: TYPE_KW@0..19 "type" [Comments("// TYPESCRIPT"), Newline("\n")] [Whitespace(" ")],
-            binding_identifier: JsIdentifierBinding {
+            binding_identifier: TsIdentifierBinding {
                 name_token: IDENT@19..21 "A" [] [Whitespace(" ")],
             },
             type_parameters: missing (required),
@@ -22,7 +22,7 @@ JsModule {
         },
         TsTypeAlias {
             type_token: TYPE_KW@29..35 "type" [Newline("\n")] [Whitespace(" ")],
-            binding_identifier: JsIdentifierBinding {
+            binding_identifier: TsIdentifierBinding {
                 name_token: IDENT@35..37 "B" [] [Whitespace(" ")],
             },
             type_parameters: missing (required),
@@ -50,7 +50,7 @@ JsModule {
         },
         TsTypeAlias {
             type_token: TYPE_KW@46..52 "type" [Newline("\n")] [Whitespace(" ")],
-            binding_identifier: JsIdentifierBinding {
+            binding_identifier: TsIdentifierBinding {
                 name_token: IDENT@52..53 "C" [] [],
             },
             type_parameters: TsTypeParameters {
@@ -103,7 +103,7 @@ JsModule {
   2: JS_MODULE_ITEM_LIST@0..81
     0: TS_TYPE_ALIAS@0..29
       0: TYPE_KW@0..19 "type" [Comments("// TYPESCRIPT"), Newline("\n")] [Whitespace(" ")]
-      1: JS_IDENTIFIER_BINDING@19..21
+      1: TS_IDENTIFIER_BINDING@19..21
         0: IDENT@19..21 "A" [] [Whitespace(" ")]
       2: (empty)
       3: EQ@21..23 "=" [] [Whitespace(" ")]
@@ -116,7 +116,7 @@ JsModule {
       5: (empty)
     1: TS_TYPE_ALIAS@29..46
       0: TYPE_KW@29..35 "type" [Newline("\n")] [Whitespace(" ")]
-      1: JS_IDENTIFIER_BINDING@35..37
+      1: TS_IDENTIFIER_BINDING@35..37
         0: IDENT@35..37 "B" [] [Whitespace(" ")]
       2: (empty)
       3: EQ@37..39 "=" [] [Whitespace(" ")]
@@ -136,7 +136,7 @@ JsModule {
       5: (empty)
     2: TS_TYPE_ALIAS@46..81
       0: TYPE_KW@46..52 "type" [Newline("\n")] [Whitespace(" ")]
-      1: JS_IDENTIFIER_BINDING@52..53
+      1: TS_IDENTIFIER_BINDING@52..53
         0: IDENT@52..53 "C" [] []
       2: TS_TYPE_PARAMETERS@53..72
         0: L_ANGLE@53..54 "<" [] []
