+++ crates/rslint_parser/test_data/inline/ok/ts_tuple_type.rast
@@ -4,7 +4,7 @@ JsModule {
     items: JsModuleItemList [
         TsTypeAlias {
             type_token: TYPE_KW@0..19 "type" [Comments("// TYPESCRIPT"), Newline("\n")] [Whitespace(" ")],
-            binding_identifier: JsIdentifierBinding {
+            binding_identifier: TsIdentifierBinding {
                 name_token: IDENT@19..21 "A" [] [Whitespace(" ")],
             },
             type_parameters: missing (required),
@@ -30,7 +30,7 @@ JsModule {
         },
         TsTypeAlias {
             type_token: TYPE_KW@44..50 "type" [Newline("\n")] [Whitespace(" ")],
-            binding_identifier: JsIdentifierBinding {
+            binding_identifier: TsIdentifierBinding {
                 name_token: IDENT@50..52 "B" [] [Whitespace(" ")],
             },
             type_parameters: missing (required),
@@ -80,7 +80,7 @@ JsModule {
         },
         TsTypeAlias {
             type_token: TYPE_KW@84..90 "type" [Newline("\n")] [Whitespace(" ")],
-            binding_identifier: JsIdentifierBinding {
+            binding_identifier: TsIdentifierBinding {
                 name_token: IDENT@90..92 "C" [] [Whitespace(" ")],
             },
             type_parameters: missing (required),
@@ -134,7 +134,7 @@ JsModule {
         },
         TsTypeAlias {
             type_token: TYPE_KW@129..135 "type" [Newline("\n")] [Whitespace(" ")],
-            binding_identifier: JsIdentifierBinding {
+            binding_identifier: TsIdentifierBinding {
                 name_token: IDENT@135..137 "D" [] [Whitespace(" ")],
             },
             type_parameters: missing (required),
@@ -160,7 +160,7 @@ JsModule {
         },
         TsTypeAlias {
             type_token: TYPE_KW@151..157 "type" [Newline("\n")] [Whitespace(" ")],
-            binding_identifier: JsIdentifierBinding {
+            binding_identifier: TsIdentifierBinding {
                 name_token: IDENT@157..159 "E" [] [Whitespace(" ")],
             },
             type_parameters: missing (required),
@@ -185,7 +185,7 @@ JsModule {
         },
         TsTypeAlias {
             type_token: TYPE_KW@174..180 "type" [Newline("\n")] [Whitespace(" ")],
-            binding_identifier: JsIdentifierBinding {
+            binding_identifier: TsIdentifierBinding {
                 name_token: IDENT@180..182 "F" [] [Whitespace(" ")],
             },
             type_parameters: missing (required),
@@ -214,7 +214,7 @@ JsModule {
   2: JS_MODULE_ITEM_LIST@0..193
     0: TS_TYPE_ALIAS@0..44
       0: TYPE_KW@0..19 "type" [Comments("// TYPESCRIPT"), Newline("\n")] [Whitespace(" ")]
-      1: JS_IDENTIFIER_BINDING@19..21
+      1: TS_IDENTIFIER_BINDING@19..21
         0: IDENT@19..21 "A" [] [Whitespace(" ")]
       2: (empty)
       3: EQ@21..23 "=" [] [Whitespace(" ")]
@@ -233,7 +233,7 @@ JsModule {
       5: (empty)
     1: TS_TYPE_ALIAS@44..84
       0: TYPE_KW@44..50 "type" [Newline("\n")] [Whitespace(" ")]
-      1: JS_IDENTIFIER_BINDING@50..52
+      1: TS_IDENTIFIER_BINDING@50..52
         0: IDENT@50..52 "B" [] [Whitespace(" ")]
       2: (empty)
       3: EQ@52..54 "=" [] [Whitespace(" ")]
@@ -270,7 +270,7 @@ JsModule {
       5: (empty)
     2: TS_TYPE_ALIAS@84..129
       0: TYPE_KW@84..90 "type" [Newline("\n")] [Whitespace(" ")]
-      1: JS_IDENTIFIER_BINDING@90..92
+      1: TS_IDENTIFIER_BINDING@90..92
         0: IDENT@90..92 "C" [] [Whitespace(" ")]
       2: (empty)
       3: EQ@92..94 "=" [] [Whitespace(" ")]
@@ -310,7 +310,7 @@ JsModule {
       5: (empty)
     3: TS_TYPE_ALIAS@129..151
       0: TYPE_KW@129..135 "type" [Newline("\n")] [Whitespace(" ")]
-      1: JS_IDENTIFIER_BINDING@135..137
+      1: TS_IDENTIFIER_BINDING@135..137
         0: IDENT@135..137 "D" [] [Whitespace(" ")]
       2: (empty)
       3: EQ@137..139 "=" [] [Whitespace(" ")]
@@ -329,7 +329,7 @@ JsModule {
       5: (empty)
     4: TS_TYPE_ALIAS@151..174
       0: TYPE_KW@151..157 "type" [Newline("\n")] [Whitespace(" ")]
-      1: JS_IDENTIFIER_BINDING@157..159
+      1: TS_IDENTIFIER_BINDING@157..159
         0: IDENT@157..159 "E" [] [Whitespace(" ")]
       2: (empty)
       3: EQ@159..161 "=" [] [Whitespace(" ")]
@@ -347,7 +347,7 @@ JsModule {
       5: (empty)
     5: TS_TYPE_ALIAS@174..193
       0: TYPE_KW@174..180 "type" [Newline("\n")] [Whitespace(" ")]
-      1: JS_IDENTIFIER_BINDING@180..182
+      1: TS_IDENTIFIER_BINDING@180..182
         0: IDENT@180..182 "F" [] [Whitespace(" ")]
       2: (empty)
       3: EQ@182..184 "=" [] [Whitespace(" ")]
