+++ crates/rslint_parser/test_data/inline/ok/ts_type_constraint_clause.rast
@@ -4,7 +4,7 @@ JsModule {
     items: JsModuleItemList [
         TsTypeAlias {
             type_token: TYPE_KW@0..19 "type" [Comments("// TYPESCRIPT"), Newline("\n")] [Whitespace(" ")],
-            binding_identifier: JsIdentifierBinding {
+            binding_identifier: TsIdentifierBinding {
                 name_token: IDENT@19..20 "A" [] [],
             },
             type_parameters: TsTypeParameters {
@@ -36,7 +36,7 @@ JsModule {
         },
         TsTypeAlias {
             type_token: TYPE_KW@43..49 "type" [Newline("\n")] [Whitespace(" ")],
-            binding_identifier: JsIdentifierBinding {
+            binding_identifier: TsIdentifierBinding {
                 name_token: IDENT@49..50 "B" [] [],
             },
             type_parameters: TsTypeParameters {
@@ -101,7 +101,7 @@ JsModule {
   2: JS_MODULE_ITEM_LIST@0..88
     0: TS_TYPE_ALIAS@0..43
       0: TYPE_KW@0..19 "type" [Comments("// TYPESCRIPT"), Newline("\n")] [Whitespace(" ")]
-      1: JS_IDENTIFIER_BINDING@19..20
+      1: TS_IDENTIFIER_BINDING@19..20
         0: IDENT@19..20 "A" [] []
       2: TS_TYPE_PARAMETERS@20..39
         0: L_ANGLE@20..21 "<" [] []
@@ -123,7 +123,7 @@ JsModule {
       5: SEMICOLON@42..43 ";" [] []
     1: TS_TYPE_ALIAS@43..88
       0: TYPE_KW@43..49 "type" [Newline("\n")] [Whitespace(" ")]
-      1: JS_IDENTIFIER_BINDING@49..50
+      1: TS_IDENTIFIER_BINDING@49..50
         0: IDENT@49..50 "B" [] []
       2: TS_TYPE_PARAMETERS@50..78
         0: L_ANGLE@50..51 "<" [] []
