+++ crates/rslint_parser/test_data/inline/ok/ts_type_operator.rast
@@ -4,7 +4,7 @@ JsModule {
     items: JsModuleItemList [
         TsTypeAlias {
             type_token: TYPE_KW@0..19 "type" [Comments("// TYPESCRIPT"), Newline("\n")] [Whitespace(" ")],
-            binding_identifier: JsIdentifierBinding {
+            binding_identifier: TsIdentifierBinding {
                 name_token: IDENT@19..21 "A" [] [Whitespace(" ")],
             },
             type_parameters: missing (required),
@@ -46,7 +46,7 @@ JsModule {
         },
         TsTypeAlias {
             type_token: TYPE_KW@48..54 "type" [Newline("\n")] [Whitespace(" ")],
-            binding_identifier: JsIdentifierBinding {
+            binding_identifier: TsIdentifierBinding {
                 name_token: IDENT@54..56 "B" [] [Whitespace(" ")],
             },
             type_parameters: missing (required),
@@ -64,7 +64,7 @@ JsModule {
         },
         TsTypeAlias {
             type_token: TYPE_KW@66..72 "type" [Newline("\n")] [Whitespace(" ")],
-            binding_identifier: JsIdentifierBinding {
+            binding_identifier: TsIdentifierBinding {
                 name_token: IDENT@72..74 "C" [] [Whitespace(" ")],
             },
             type_parameters: missing (required),
@@ -131,7 +131,7 @@ JsModule {
   2: JS_MODULE_ITEM_LIST@0..129
     0: TS_TYPE_ALIAS@0..48
       0: TYPE_KW@0..19 "type" [Comments("// TYPESCRIPT"), Newline("\n")] [Whitespace(" ")]
-      1: JS_IDENTIFIER_BINDING@19..21
+      1: TS_IDENTIFIER_BINDING@19..21
         0: IDENT@19..21 "A" [] [Whitespace(" ")]
       2: (empty)
       3: EQ@21..23 "=" [] [Whitespace(" ")]
@@ -161,7 +161,7 @@ JsModule {
       5: SEMICOLON@47..48 ";" [] []
     1: TS_TYPE_ALIAS@48..66
       0: TYPE_KW@48..54 "type" [Newline("\n")] [Whitespace(" ")]
-      1: JS_IDENTIFIER_BINDING@54..56
+      1: TS_IDENTIFIER_BINDING@54..56
         0: IDENT@54..56 "B" [] [Whitespace(" ")]
       2: (empty)
       3: EQ@56..58 "=" [] [Whitespace(" ")]
@@ -174,7 +174,7 @@ JsModule {
       5: SEMICOLON@65..66 ";" [] []
     2: TS_TYPE_ALIAS@66..94
       0: TYPE_KW@66..72 "type" [Newline("\n")] [Whitespace(" ")]
-      1: JS_IDENTIFIER_BINDING@72..74
+      1: TS_IDENTIFIER_BINDING@72..74
         0: IDENT@72..74 "C" [] [Whitespace(" ")]
       2: (empty)
       3: EQ@74..76 "=" [] [Whitespace(" ")]
