+++ crates/rslint_parser/test_data/inline/ok/ts_type_parameters.rast
@@ -4,7 +4,7 @@ JsModule {
     items: JsModuleItemList [
         TsTypeAlias {
             type_token: TYPE_KW@0..19 "type" [Comments("// TYPESCRIPT"), Newline("\n")] [Whitespace(" ")],
-            binding_identifier: JsIdentifierBinding {
+            binding_identifier: TsIdentifierBinding {
                 name_token: IDENT@19..20 "A" [] [],
             },
             type_parameters: TsTypeParameters {
@@ -134,7 +134,7 @@ JsModule {
   2: JS_MODULE_ITEM_LIST@0..109
     0: TS_TYPE_ALIAS@0..109
       0: TYPE_KW@0..19 "type" [Comments("// TYPESCRIPT"), Newline("\n")] [Whitespace(" ")]
-      1: JS_IDENTIFIER_BINDING@19..20
+      1: TS_IDENTIFIER_BINDING@19..20
         0: IDENT@19..20 "A" [] []
       2: TS_TYPE_PARAMETERS@20..87
         0: L_ANGLE@20..21 "<" [] []
