+++ crates/rslint_parser/test_data/inline/ok/ts_type_predicate.rast
@@ -4,7 +4,7 @@ JsModule {
     items: JsModuleItemList [
         TsTypeAlias {
             type_token: TYPE_KW@0..19 "type" [Comments("// TYPESCRIPT"), Newline("\n")] [Whitespace(" ")],
-            binding_identifier: JsIdentifierBinding {
+            binding_identifier: TsIdentifierBinding {
                 name_token: IDENT@19..21 "A" [] [Whitespace(" ")],
             },
             type_parameters: missing (required),
@@ -40,7 +40,7 @@ JsModule {
         },
         TsTypeAlias {
             type_token: TYPE_KW@42..48 "type" [Newline("\n")] [Whitespace(" ")],
-            binding_identifier: JsIdentifierBinding {
+            binding_identifier: TsIdentifierBinding {
                 name_token: IDENT@48..50 "B" [] [Whitespace(" ")],
             },
             type_parameters: missing (required),
@@ -76,7 +76,7 @@ JsModule {
         },
         TsTypeAlias {
             type_token: TYPE_KW@79..85 "type" [Newline("\n")] [Whitespace(" ")],
-            binding_identifier: JsIdentifierBinding {
+            binding_identifier: TsIdentifierBinding {
                 name_token: IDENT@85..93 "asserts" [] [Whitespace(" ")],
             },
             type_parameters: missing (required),
@@ -88,7 +88,7 @@ JsModule {
         },
         TsTypeAlias {
             type_token: TYPE_KW@102..108 "type" [Newline("\n")] [Whitespace(" ")],
-            binding_identifier: JsIdentifierBinding {
+            binding_identifier: TsIdentifierBinding {
                 name_token: IDENT@108..110 "C" [] [Whitespace(" ")],
             },
             type_parameters: missing (required),
@@ -120,7 +120,7 @@ JsModule {
   2: JS_MODULE_ITEM_LIST@0..126
     0: TS_TYPE_ALIAS@0..42
       0: TYPE_KW@0..19 "type" [Comments("// TYPESCRIPT"), Newline("\n")] [Whitespace(" ")]
-      1: JS_IDENTIFIER_BINDING@19..21
+      1: TS_IDENTIFIER_BINDING@19..21
         0: IDENT@19..21 "A" [] [Whitespace(" ")]
       2: (empty)
       3: EQ@21..23 "=" [] [Whitespace(" ")]
@@ -146,7 +146,7 @@ JsModule {
       5: SEMICOLON@41..42 ";" [] []
     1: TS_TYPE_ALIAS@42..79
       0: TYPE_KW@42..48 "type" [Newline("\n")] [Whitespace(" ")]
-      1: JS_IDENTIFIER_BINDING@48..50
+      1: TS_IDENTIFIER_BINDING@48..50
         0: IDENT@48..50 "B" [] [Whitespace(" ")]
       2: (empty)
       3: EQ@50..52 "=" [] [Whitespace(" ")]
@@ -172,7 +172,7 @@ JsModule {
       5: SEMICOLON@78..79 ";" [] []
     2: TS_TYPE_ALIAS@79..102
       0: TYPE_KW@79..85 "type" [Newline("\n")] [Whitespace(" ")]
-      1: JS_IDENTIFIER_BINDING@85..93
+      1: TS_IDENTIFIER_BINDING@85..93
         0: IDENT@85..93 "asserts" [] [Whitespace(" ")]
       2: (empty)
       3: EQ@93..95 "=" [] [Whitespace(" ")]
@@ -181,7 +181,7 @@ JsModule {
       5: SEMICOLON@101..102 ";" [] []
     3: TS_TYPE_ALIAS@102..126
       0: TYPE_KW@102..108 "type" [Newline("\n")] [Whitespace(" ")]
-      1: JS_IDENTIFIER_BINDING@108..110
+      1: TS_IDENTIFIER_BINDING@108..110
         0: IDENT@108..110 "C" [] [Whitespace(" ")]
       2: (empty)
       3: EQ@110..112 "=" [] [Whitespace(" ")]
