+++ crates/rslint_syntax/src/generated.rs
@@ -333,6 +333,7 @@ pub enum JsSyntaxKind {
     JS_EXPORT_AS_CLAUSE,
     JS_LITERAL_EXPORT_NAME,
     JS_AWAIT_EXPRESSION,
+    TS_IDENTIFIER_BINDING,
     TS_ANY_TYPE,
     TS_UNKNOWN_TYPE,
     TS_NUMBER_TYPE,
